app/requirements.txt
astroid==2.3.3
certifi==2019.9.11
Click==7.0
Flask==1.1.1
Flask-SocketIO==4.2.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.3
olefile==0.46
Pillow==6.2.1
pylint==2.4.4
python-engineio==3.9.3
python-socketio==4.3.1
six==1.13.0
Werkzeug==0.16.0
wrapt==1.11.2
